Add hidden-player fallback assignment and shared editor
