+++ src/lib/stores/player_store.ts
@@ -49,10 +49,7 @@ function normalizePositions(positions: PlayerPosition[]): PlayerPosition[] {
 	const normalized = positions.map((pos) => ({
 		...pos,
 		role: pos.role ?? "secondary",
-		skill:
-			pos.skill === "medium"
-				? "mid"
-				: pos.skill ?? "mid",
+		skill: pos.skill ?? "mid",
 	}));
 
 	let primary_index = normalized.findIndex((pos) => pos.role === "primary");
